Refs #27418 - Smart Proxy Pulp based registry URL
This patch uses the Smart Proxy Pulp setting registry_url to determine
where it can find the container registry. It assumes the same CA is used
as the Smart Proxy itself rather than relying on the hard coded config.

If there's a need for a different value, this could also be stored as a
config and learned dynamically via a trusted channel that is the Proxy
Registration.