Fixes #30368 - Change Candlepin and Pulp 3 CA handling
This change makes the Candlepin CA file optional by falling back to the
Foreman CA file. The result is a reduced configuration in most
deployments.

For Pulp 3 the certificate is now read in the same way as Candlepin.
Previously it relies on the CA being in the global allowed CA. This may
be an issue in some cases, but in the default deployment it isn't.
Following the general SSL config makes the configuration more
predictable for users. It can also be easier in a containerized setup or
on a system where the admin is not allowed to modify the system CA
certificates.

In the ping model the verify_ssl option is dropped. This isn't respected
elsewhere and it's misleading to have a valid ping only to have it fail
at runtime.

It also changes Crane's setting from registry_ca_cert_file to
crane_ca_cert_file. This follows what's in the example config, unit
tests and installer module. The fact it was not working suggests that it
was using the system CA store.