Fixes #32835 - Add RHSM and content URL as info providers
This introduces two Smart Proxy extension methods and exposes them as
info providers. This allows clients to use the ENC API to (re)configure
subscription-manager.

First of all, this implements logic to determine the RHSM URL. This is
constructed by looking at Pulp. While this is really the wrong way,
there is currently no alternative.

Ideally this would be exposed as a setting as well. There is a plan to
align everything on port 443. If the Smart Proxy exposes a setting,
Katello doesn't need to guess and a smooth migration can be realized.

This all becomes more important when the host subscription RPM is phased
out.

Then it also introduces pulp_content_url which is already a setting
that's exposed via the Smart Proxy Pulp feature. This is also used in
the various places where it's needed.

It does highlight a missing setting for pulp_ansible content. Again,
Katello should not have to guess URLs but discover it from the
infrastructure.

Another potential use case that this allows is using REX to move a host
from one content source to another.